Moves decrypt methods to module
